# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/recommendations/combo-suggestions": {
      "get": {
        "operationId": "RestaurantController_getComboSuggestions",
        "parameters": [
          {
            "name": "dishId",
            "required": true,
            "in": "query",
            "description": "ID món ăn gốc",
            "schema": {
              "example": "item_1_1_26076",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Số lượng gợi ý",
            "schema": {
              "default": 3,
              "example": 3,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Restaurants"
        ]
      }
    },
    "/api/v1/search": {
      "get": {
        "operationId": "RestaurantController_searchRestaurants",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "description": "Từ khóa tìm kiếm",
            "schema": {
              "example": "pizza",
              "type": "string"
            }
          },
          {
            "name": "customerId",
            "required": false,
            "in": "query",
            "schema": {
              "example": "cust_1721973133_00006",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Trang kết quả (bắt đầu từ 1)",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Restaurants"
        ]
      }
    },
    "/api/v1/restaurant/{id}/profile": {
      "get": {
        "operationId": "RestaurantController_getRestaurantProfile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID của nhà hàng",
            "schema": {
              "example": "123",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy thông tin cửa hàng thành công."
          }
        },
        "summary": "Lấy thông tin hồ sơ nhà hàng",
        "tags": [
          "Restaurants"
        ]
      }
    },
    "/api/v1/restaurant/{id}/menu": {
      "get": {
        "operationId": "RestaurantController_getRestaurantMenu",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID của nhà hàng",
            "schema": {
              "example": "123",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy thông tin thực đơn thành công."
          }
        },
        "summary": "Lấy danh sách thực đơn của nhà hàng",
        "tags": [
          "Restaurants"
        ]
      }
    },
    "/api/v1/restaurant/{id}/top-dishes": {
      "get": {
        "operationId": "RestaurantController_getTopDishes",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID của nhà hàng",
            "schema": {
              "example": "123",
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": false,
            "in": "query",
            "description": "Ngày cần xem (YYYY-MM-DD)",
            "schema": {
              "example": "2026-06-26",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy danh sách top món ăn bán chạy thành công."
          }
        },
        "summary": "Lấy top món ăn bán chạy trong ngày của nhà hàng",
        "tags": [
          "Restaurants"
        ]
      }
    },
    "/api/v1/customers/{customerId}/orders/history": {
      "get": {
        "operationId": "CustomerController_getOrderHistory",
        "parameters": [
          {
            "name": "customerId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "cust_123",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "default": "desc",
              "example": "asc",
              "type": "string"
            }
          },
          {
            "name": "lastOrderedAt",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2026-01-01T00:00:00.000Z",
              "type": "string"
            }
          },
          {
            "name": "lastOrderId",
            "required": false,
            "in": "query",
            "schema": {
              "example": "item_1_1_26076",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Customer",
          "Customer"
        ]
      }
    }
  },
  "info": {
    "title": "Restaurant API",
    "description": "API documentation",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {}
  }
}